Move policy export settings from model to policy class
