+++ library/src/physicalai/policies/act/policy.py
@@ -8,6 +8,13 @@ from typing import Any, cast
 import torch
 
 from physicalai.data import Dataset, Feature, FeatureType, NormalizationParameters, Observation
+from physicalai.export.backends import (
+    ExecuTorchExportParameters,
+    ExportParameters,
+    ONNXExportParameters,
+    OpenVINOExportParameters,
+    TorchExportParameters,
+)
 from physicalai.export.mixin_policy import ExportablePolicyMixin, ExportBackend
 from physicalai.gyms import Gym
 from physicalai.policies.act.config import ACTConfig
@@ -440,3 +447,27 @@ class ACT(ExportablePolicyMixin, Policy):
             ExportBackend.ONNX,
             ExportBackend.EXECUTORCH,
         ]
+
+    @property
+    def extra_export_args(self) -> dict[str, ExportParameters]:
+        """Additional export arguments for model conversion.
+
+        Returns:
+            dict[str, ExportParameters]: A dictionary mapping format names to their export parameters.
+        """
+        extra_args: dict[str, ExportParameters] = {}
+        extra_args["onnx"] = ONNXExportParameters(
+            exporter_kwargs={
+                "output_names": ["action"],
+            },
+        )
+        extra_args["openvino"] = OpenVINOExportParameters(
+            outputs=["action"],
+            export_tokenizer=False,
+            compress_to_fp16=False,
+            exporter_kwargs={},
+        )
+        extra_args["executorch"] = ExecuTorchExportParameters()
+        extra_args["torch"] = TorchExportParameters()
+
+        return extra_args
