+++ library/tests/unit/export/test_executorch_integration.py
@@ -43,8 +43,6 @@ class _ExportWrapper(ExportablePolicyMixin):
 
     def __init__(self, model: torch.nn.Module) -> None:
         self.model = model
-        if not hasattr(model, "extra_export_args"):
-            model.extra_export_args = {}
 
     @property
     def metadata_extra(self) -> dict[str, Any]:
